Display thumbnails only for calls that draw something
Rather than accumulate all thumbnails that attempt to be drawn, this patch
filters out all calls that don't draw anything.

The calls that draw selected are a superset of the calls that
SettingsDialog initializes and which is displayed by the dialog box opened
with "Trace | Options". There, the following calls are selected:
     glDraw
     glVertex
     glBegin
     glEnd

To this collection of calls, this patch also selects thumbnails for:
   glClear
   glCallList
since the latter may result in drawing by a display list.

To accomplish this filtering, the method isDrawCall() was added to
the class ApiTraceCall. This code is modelled upon similar code found
in ApiTraceFilter.